Moving pragma function into interceptors + tests
